#112 Move provider endpoints to be outside API endpoints and prefix with `providers`.
