+++ src/providers/routes.ts
@@ -1,0 +1,9 @@
+import express from 'express'
+import Controller from './controller'
+
+const router = express.Router()
+
+router.get('/connect/:provider', Controller.connect)
+router.get('/callback/:provider', Controller.callback)
+
+export default router
